Create README.md
CAN is extensively used in automotive but it has found applications everywhere. There are many “application” layers available for CAN such as ISO 15765 (cars), J1939 (trucks), DeviceNET and CANopen (both are for factory automation) but it is very easy to develop your own protocol that will fit and simplify your needs. Modern CAN transceivers provide a stable and reliable CAN physical environment without the need for expensive coaxial cables. Nearly all of the mystery of CAN has dissipated over the years. There is plenty of example CAN software to help you develop your own network.
Many think CAN is just for automotive, but this is not true. CAN has become the standard for vehicle networks, but it has been adopted in most other fields. As you find out in these pages, there are no attributes in the Bosch CAN specification that are automotive related. It is completely generic. You can easily implement your own protocol on top of CAN.
A CAN controller is a sophisticated device. Nearly all the features of the CAN protocol described here are automatically handled by the controller with almost no intervention by the host processor. All you need to do in practice is to configure the controller by writing to its
registers, write data to the controller and the controller then does all the housekeeping work to get your message on the bus.
The controller will read any frames it sees on the bus and hold them in a small FIFO memory. It will notify the host processor that this
data is available which it then reads from the controller.
The controller also contains a hardware filter mechanism that can be programmed to ignore and discard those CAN frames you do not
want passed to the processor. This saves on processor overhead.